RecordsResultsListItem: refactor and add overridable
