+++ invenio_app_rdm/theme/assets/semantic-ui/js/invenio_app_rdm/frontpage/index.js
@@ -7,12 +7,16 @@
 import React from "react";
 import ReactDOM from "react-dom";
 import RecordsList from "./RecordsList";
+import { OverridableContext } from "react-overridable";
+import { overriddenComponents } from "./override";
 
 const recordsListContainer = document.getElementById("records-list");
 const title = recordsListContainer.dataset.title;
 const fetchUrl = recordsListContainer.dataset.fetchUrl;
 
 ReactDOM.render(
-  <RecordsList title={title} fetchUrl={fetchUrl} />,
+  <OverridableContext.Provider value={overriddenComponents}>
+    <RecordsList title={title} fetchUrl={fetchUrl} />
+  </OverridableContext.Provider>,
   recordsListContainer
 );
