+++ invenio_app_rdm/theme/assets/semantic-ui/js/invenio_app_rdm/frontpage/override.js
@@ -1,0 +1,7 @@
+// This file is part of InvenioRDM
+// Copyright (C) 2022 CERN.
+//
+// Invenio App RDM is free software; you can redistribute it and/or modify it
+// under the terms of the MIT License; see LICENSE file for more details.
+
+export const overriddenComponents = {};
