+++ invenio_app_rdm/theme/assets/semantic-ui/js/invenio_app_rdm/search/components.js
@@ -25,108 +25,16 @@ import {
   Header,
   Icon,
   Input,
-  Item,
   Label,
   Message,
   Segment,
 } from "semantic-ui-react";
-import { SearchItemCreators } from "../utils";
+import RecordsResultsListItem from "../components/RecordsResultsListItem";
 import PropTypes from "prop-types";
-
-export const RDMRecordResultsListItem = ({ currentQueryState, result }) => {
-  const accessStatusId = _get(result, "ui.access_status.id", "open");
-  const accessStatus = _get(result, "ui.access_status.title_l10n", "Open");
-  const accessStatusIcon = _get(result, "ui.access_status.icon", "unlock");
-  const createdDate = _get(
-    result,
-    "ui.created_date_l10n_long",
-    "No creation date found."
-  );
-  const creators = result.ui.creators.creators.slice(0, 3);
-
-  const descriptionStripped = _get(result, "ui.description_stripped", "No description");
-
-  const publicationDate = _get(
-    result,
-    "ui.publication_date_l10n_long",
-    "No publication date found."
-  );
-  const resourceType = _get(result, "ui.resource_type.title_l10n", "No resource type");
-  const subjects = _get(result, "ui.subjects", []);
-  const title = _get(result, "metadata.title", "No title");
-  const version = _get(result, "ui.version", null);
-  const versions = _get(result, "versions");
-
-  const filters = currentQueryState && Object.fromEntries(currentQueryState.filters);
-  const allVersionsVisible = filters?.allversions;
-  const numOtherVersions = versions.index - 1;
-  // Derivatives
-  const viewLink = `/records/${result.id}`;
-  return (
-    <Item>
-      <Item.Content>
-        <Item.Extra className="labels-actions">
-          <Label size="tiny" className="primary">
-            {publicationDate} ({version})
-          </Label>
-          <Label size="tiny" className="neutral">
-            {resourceType}
-          </Label>
-          <Label size="tiny" className={`access-status ${accessStatusId}`}>
-            {accessStatusIcon && <i className={`icon ${accessStatusIcon}`} />}
-            {accessStatus}
-          </Label>
-        </Item.Extra>
-        <Item.Header as="h2">
-          <a href={viewLink}>{title}</a>
-        </Item.Header>
-        <Item className="creatibutors">
-          <SearchItemCreators creators={creators} />
-        </Item>
-        <Item.Description>
-          {_truncate(descriptionStripped, { length: 350 })}
-        </Item.Description>
-        <Item.Extra>
-          {subjects.map((subject) => (
-            <Label key={subject.title_l10n} size="tiny">
-              {subject.title_l10n}
-            </Label>
-          ))}
-          {createdDate && (
-            <p>
-              <small>
-                {i18next.t("Uploaded on")} <span>{createdDate}</span>
-              </small>
-            </p>
-          )}
-          {!allVersionsVisible && versions.index > 1 && (
-            <p>
-              <small>
-                <b>
-                  {numOtherVersions} more{" "}
-                  {numOtherVersions > 1 ? "versions" : "version"} exist for this record
-                </b>
-              </small>
-            </p>
-          )}
-        </Item.Extra>
-      </Item.Content>
-    </Item>
-  );
-};
-
-RDMRecordResultsListItem.propTypes = {
-  currentQueryState: PropTypes.object,
-  result: PropTypes.object.isRequired,
-};
-
-RDMRecordResultsListItem.defaultProps = {
-  currentQueryState: null,
-};
 
 export const RDMRecordResultsListItemWithState = withState(
   ({ currentQueryState, result }) => (
-    <RDMRecordResultsListItem currentQueryState={currentQueryState} result={result} />
+    <RecordsResultsListItem currentQueryState={currentQueryState} result={result} />
   )
 );
 
@@ -135,7 +43,7 @@ RDMRecordResultsListItemWithState.propTypes = {
   result: PropTypes.object.isRequired,
 };
 
-RDMRecordResultsListItem.defaultProps = {
+RDMRecordResultsListItemWithState.defaultProps = {
   currentQueryState: null,
 };
